Solución de dos warnings

+++ DeathTokenBase/DeathToken/UI.cpp
@@ -54,12 +54,12 @@ UI::UI(GameState* gS, Game* game) : gS(gS), game(game), onBet(false), chipOnUse(
 inline int
 UI::relativeX(const float& n)
 {
-	return (n / 1920.0f) * Game::WIN_WIDTH;
+	return (int)((n / 1920.0f) * Game::WIN_WIDTH);
 }
 inline int
 UI::relativeY(const float& n)
 {
-	return (n / 1080.0f) * Game::WIN_HEIGHT;
+	return (int)((n / 1080.0f) * Game::WIN_HEIGHT);
 }
 void
 UI::OnExit()
